Delete folder

+++ python/deleteFolderInPython.py
@@ -32,6 +32,7 @@ def checkOs():
         operatingSystem = "Linux"
 
     print("Finished checking operating system at", datetime.now().strftime("%m-%d-%Y %I:%M %p"))
+
     print("")
     return operatingSystem
 
@@ -84,10 +85,7 @@ def checkParameters(pathToFolder, folderName):
         print("")
 
     else: 
-        print(Fore.RED + "One or more parameter(s) are incorrect." + Style.RESET_ALL)
-
-        print("Finished checking parameter(s) at", datetime.now().strftime("%m-%d-%Y %I:%M %p"))
-        exit("")
+        raise Exception("One or more parameter(s) are incorrect.")
 
 
 def deleteFolder(): 
@@ -110,7 +108,11 @@ def deleteFolder():
         print("Started deleting {0} folder at {1}".format(folderName, startDateTime.strftime("%m-%d-%Y %I:%M %p")))
 
         deletePath="{0}/{1}".format(pathToFolder, folderName)
-        os.rmdir(deletePath)
+        
+        if os.path.isdir(deletePath):
+            os.rmdir(deletePath)
+        else: 
+            raise Exception("Folder path is invalid.")
 
         print(Fore.GREEN + "Successfully deleted {0} folder.".format(folderName) + Style.RESET_ALL)
